requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-cloudinary-storage==0.3.0
cloudinary==1.40.0
django-filter==24.3
dj-database-url==2.2.0
drf-spectacular==0.27.2
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
whitenoise==6.11.0
Pillow==12.0.0
